Add station position and func that renames routes
Renamed column station_index to position in railway_stations_routes.
In route#show added feature to setting station position on viewed route,
by creating new action - update_position in railway_stations_controller.
Added a callback in the route, calling a function that sets
the name of the route by the first and last stations in the list.